can choose exist tags on compose page
