+++ backend/src/resource/asset/index.html
@@ -10,6 +10,6 @@
             .tag-btn{color:white;border-radius:4px;text-shadow:0 1px 1px rgba(0, 0, 0, 0.2);background:rgb(28, 184, 65);margin-right:6px}
             .tag-link{margin-right:6px}
         </style>
-    <script type="module">import init from '/asset/index-f1b1ef641399ddf5.js';init('/asset/index-f1b1ef641399ddf5_bg.wasm');</script></head>
+    <script type="module">import init from '/asset/index-d1a376e8d6448cde.js';init('/asset/index-d1a376e8d6448cde_bg.wasm');</script></head>
     <body>
 </body></html>
