+++ backend/src/resource/asset/snippets/blog-frontend-3607fd663da6040e/asset/editor.js
@@ -35,8 +35,8 @@ export function inputTag(event) {
     tagInput.focus();
 }
 
-export function selectTag(tagBox) {
-    addTag(tagBox.childNodes[0].nodeValue);
+export function selectTag(tag) {
+    addTag(tag);
 }
 
 function addTag(val) {
