+++ backend/src/service/asset_list.rs
@@ -1,7 +1,7 @@
 [
 ("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
-("index-f1b1ef641399ddf5.js", include_bytes!(r#"..\resource\asset/index-f1b1ef641399ddf5.js.gz"#), "text/javascript"),
-("index-f1b1ef641399ddf5_bg.wasm", include_bytes!(r#"..\resource\asset/index-f1b1ef641399ddf5_bg.wasm.gz"#), "application/wasm"),
+("index-d1a376e8d6448cde.js", include_bytes!(r#"..\resource\asset/index-d1a376e8d6448cde.js.gz"#), "text/javascript"),
+("index-d1a376e8d6448cde_bg.wasm", include_bytes!(r#"..\resource\asset/index-d1a376e8d6448cde_bg.wasm.gz"#), "application/wasm"),
 ("index.html", include_bytes!(r#"..\resource\asset/index.html.gz"#), "text/html; charset=utf-8"),
 ("purecss.min-b565cdb90b7aac9f.css", include_bytes!(r#"..\resource\asset/purecss.min-b565cdb90b7aac9f.css.gz"#), "text/css"),
 ("snippets/blog-frontend-3607fd663da6040e/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-3607fd663da6040e\asset\editor.js.gz"#), "text/javascript"),
